requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# OCR Document Translator — Python dependencies
# Install with:  pip install -r requirements.txt
# ──────────────────────────────────────────────────────────────────────────────

# UI framework
streamlit>=1.32.0

# OCR
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0
opencv-python>=4.8.0

# PDF parsing and image extraction
pymupdf>=1.23.0          # imports as 'fitz'

# PDF reconstruction
reportlab>=4.1.0

# DOCX reconstruction
python-docx>=1.1.0

# Offline translation (Russian → English)
argostranslate>=1.9.0

# Numerical / data
numpy>=1.24.0
pandas>=2.0.0
